requirements.txt
langchain
langchain-core
langchain-openai
langchain_deepseek
langchain-anthropic
langchain-community
langchain-google-genai
langchain-ollama
langgraph
chainlit
plotly
kaleido
openai-whisper
ffmpeg-python